% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Align.Concat.R
\name{Align.Concat}
\alias{Align.Concat}
\title{Concatenate alignments from different gene regions into a supermatrix at the species level}
\usage{
Align.Concat(input = NULL, Sp.List.NoDNA = NULL, outputConcat = NULL)
}
\arguments{
\item{input}{the path to the folder storing the alignments (alignments have to be in}

\item{Sp.List.NoDNA}{a vector of the species without DNA that should be included
in the alignment, or the option can be NULL (or left blank),
in which case the function automatically creates a complete
 species list of all the species present in the different alignments.}

\item{outputConcat}{Name of the supermatrix (can include the path as well).}
}
\value{
The function returns 1) the differents alignments
separately in nexus format .nex, including taxa that do not have DNA sequence
information (replace by '----') (all these files can be loaded separetely in
BEAST for instance), 2) a concatenation (a supermatrix) of all the sequences of
the different gene regions in nexus and fasta format, 3) a partition file in
txt format including the partitions of the different gene regions in the
concatenated file, this file is designed to be RAxML compatible (RAxML manual
v8.2, https://sco.h-its.org/exelixis/resource/download/NewManual.pdf).
}
\description{
This function concatenates the alignments from different gene regions in a
single supermatrix in nexus and fasta formats, at the species level.  The function also allows the
inclusion of species without DNA, if necessary (using BEAST to resolve
polytomies for instance).
}
\examples{
# Run the function to build a supermatrix
\dontrun{

# To run the example it might be better to copy the input alignment files
# provided by the package to a temporary directory created into the
# current working directory.
src.dir = system.file("extdata/multi.align/ForConcat", package = "regPhylo")
dir.create("TempDir.ForConcat")
# Set up the path of the TempDir folder.
dest.dir = paste(getwd(), "/TempDir.ForConcat", sep="")
file.names <- dir(src.dir)
# Copy all the files stored in regPhylo/extdata/multi.align/ForConcat"
# into a temporary folder.
sapply(file.names, function(x) {
file.copy(from = paste(src.dir, x, sep = "/"),
to = paste(dest.dir, x, sep = "/"),
overwrite = FALSE) })

# Run the function to build the supermatrix.
Align.Concat(input = "TempDir.ForConcat", Sp.List = NULL, outputConcat = NULL)

#' # Import the supermatrix in R
require(ape)
Supermatrix = read.dna("TempDir.ForConcat/Concat.fas", format = "fasta")

# Create another temporary file to build a supermatrix including species without DNA.
dir.create("TempDir.ForConcat2")
file.names <- dir("TempDir.ForConcat")
file.names <- file.names[grep(".fas", file.names)] # select only the .fas alignment
file.names <- file.names[-grep("Concat.fas", file.names)] # remove the Concat.fas alignment just created above.
sapply(file.names, function(x) {
file.copy(from = paste("TempDir.ForConcat", x, sep = "/"),
to = paste("TempDir.ForConcat2", x, sep = "/"),
overwrite = FALSE) })


# Run the function to build a supermatrix including two species without DNA.
Align.Concat(input = "TempDir.ForConcat2",
Sp.List = c("Titi_titi", "Toto_toto"),
outputConcat = "TempDir.ForConcat2/Concat_2spNoDNA")

# Import the supermatrix in R
Supermatrix2SpNoDNA = read.dna("TempDir.ForConcat2/Concat_2spNoDNA.fas", format = "fasta")


# To clean the files created while running the example do the following:
unlink("TempDir.ForConcat", recursive = TRUE)
unlink("TempDir.ForConcat2", recursive = TRUE)

}

}
