% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{alignment2matrix}
\alias{alignment2matrix}
\title{Convert an alignment object (seqinr format) to a matrix}
\usage{
alignment2matrix(input)
}
\arguments{
\item{input}{an alignment object (seqinr format)}
}
\description{
This function converts an alignment object (seqinr format) to a matrix, when the initial 
alignment was converted from a DNAbin object loaded with the ape::read.FASTA function.
It solves some recent probelm I encounter with the as.matrix.alignment function 
from the seqinr package
}
