% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmMissData.R
\name{AmMissData}
\alias{AmMissData}
\title{Estimate the percentage of missing gene region within a species-by-gene matrix}
\usage{
AmMissData(input = NULL, gene.list = NULL)
}
\arguments{
\item{input}{the species-by-gene matrix which is provided as the first object in the
list returned by the function SpeciesGeneMat.Bl.R.}

\item{gene.list}{a vector of the selected gene regions (gene names
have to be consistent with the header of the table with the suffix
'_CleanDataset.txt' exported by the function SpeciesGeneMat_Bl.R).
Example (Cytochrome c oxydase subunit 1, should be written 'co1' and not 'COI' or 'COX1').}
}
\value{
The function return the percentage of missing gene regions.
}
\description{
This function extracts the percentage of missing data in the species-by-gene
matrix according to a list of selected gene regions.
}
\examples{
# Load a species-by-gene matrix for instance exported by SpeciesGeneMat.Bl function.
data(Seq.DF4) ## the first object of the list is the Species-by-gene matrix

# Run the function
AmMissData(input = Seq.DF4[[1]], gene.list = c("co1", "16srrna"))

}
