% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{Clean.Seq}
\alias{Clean.Seq}
\title{Help to detect outlier sequences in an alignment and remove them.}
\usage{
Clean.Seq(
  inputal = NULL,
  Outlier.Th = 1.5,
  output = NULL,
  Evol.Model = "TN93",
  Sum.Stat = "median",
  Iter.optim = TRUE,
  Re.Align = FALSE,
  nbthreads = 1,
  Approxim = FALSE,
  nb.seq = 25,
  nb.rep = 5
)
}
\arguments{
\item{inputal}{an object of class "alignment" (seqinr R package),
or "DNAbin" (ape R package) or the name (including the path if necessary) of the input fasta file. The input file must be a DNA alignement file already aligned.}

\item{Outlier.Th}{is a number (by default 1.5) defining the distance to consider an outlier sequences. The Outlier.Th is a multiplier of the interquartile range from the genetic distances distribution using the Sum.stat (eg. median). This parameter is equivalent of the "range" parameter from the boxplot function from the stat R package.}

\item{output}{name of the two alignment exported. The name can also include the path the folder.
Two separate alignment are exported in a fasta format, the alignment with the ".Clean.haplo.fas" siffix includes only the haplotypes after removing outlier sequences, the alignment with the ".Clean.AllSeq.fas" suffix includes all the sequences after removing the outlier sequences.}

\item{Evol.Model}{evolutionary model used to compute the DNA distance among sequences, default "TN93" (Tamura and Nei 1993), This options accepts all the models used in dist.dna function from the ape R package.}

\item{Sum.Stat}{type of symmary statistic ("median" by default)  to compare the DNA distance of the focal sequence with the other, by defult we use the median distance but it can also be the "min" for the minimal dNA distance between 1 sequence and its closest neighbor.}

\item{Iter.optim}{if TRUE then an iterative outlier detection appraoch is performed until no longer outlier sequenece are detected considering the outier threshold defined by the Outlier.Th option.}

\item{Re.Align}{the sequences are realigned after the indel gaps were removed using fast mafft.fftns1 algorithm, but this steps can be long especially combined the automatic iteration of the optimisation approach that can be perfomed multiple times.}

\item{nbthreads}{a number (by default 1) defining the number of threads used to perfoimed the Approximate detection of outlier sequences}

\item{Approxim}{if TRUE then the function perform an approximate estimate of the outlier sequence detection by using a random selection of nb.seq (25 by default) sequences nb.rep (5 by default) times for each sequence, this options is worth using for large alignement including more than 10 000 sequences.}

\item{nb.seq}{is a number defining how many sequence need to be selected randomly if Approxim is TRUE, increasing the number of sequence will considerably slow dow the algorithm. default 25.}

\item{nb.rep}{is a number defining how many replicate of the randomw selection of sequence must be performed for each sequence of interest (default 5 times)}
}
\value{
The function return a list of 4 objects using the following headers:
\itemize{
\item 'Outlier.Sequences' which provides a vectorwith teh names of all the sequences considered as potential outlier.
\item 'Clean.Haplo.Alignment' is a DNAbin object of the alignment of only the haplotypes sequences after removing the outlier sequences.
\item 'Clean.All.Seq.Alignment' is a DNAbin object of the alignment of all the sequences after removing the outlier sequences.
\item 'First.Dist.Mat' is the summary statistic (either median or minimal distance, this parameter is defined by the 'Sum.Stat' option) of the genetic distance from the each focal sequences in comparison to all the other.
}
}
\description{
This function helps to detect some potential outlier sequences
(i.e. mis-aligned sequences, which might be caused by different problems such
as gene or species annotation problems, presence of paralogous sequences, etc)
that should be removed from the pool of sequences (per species and gene regions)
before selecting the best sequence.  This function is only provided as a tool
to HELP detect potential outlier sequences, great care must be taken to
ensure sequences are not erroneously omitted as potential outliers,
or remain included undetected outlier sequences. We recommend checking
alignments by eye to ensure your confidence in the retained sequences.
}
\details{
This function uses the following general strategy to detect outlier
sequences.  First, the function detects and removes the duplicated sequences to work considering only the haplotypes. Second, a distance matrix between all pairwise sequences is built considering different strategies implemented in the dist.DNA function from the ape R package.
Third the function computes a summary statistic (can be the "median" or the "min" for the minimal distance) for each focal sequence with the genetic distance with all the other sequences using the Sum.Stat option. Fourth, the distribution of summary statistic is compared to an outlier threshold defined by the Outlier.Th option. Outlier.Th is a multiplier of the interquartile range distance as defined by the "range" parameter from the boxplot R function. All the sequences with a summary statistic above the threshold are considered as potential outliers and are removed from the final alignment. Then, the deletion gap only sites are removed automatically and the retained sequences can be re-aligned using a MAFFT FFT NS1 algorithm if desired.
In order to speed up the detection of outlier sequences, we also provide an approximate option to compute the pairwise genetic distances among sequences. The option "Approxim = TRUE" allows to perform such approximative detection by randomly selecting a number of sequences (set by the option nb.seq) several times (set by the option nb.rep) to compare the average summary statistic of each focal sequences. We can here higly parallelized this approximate search. By default, the function selects 25 random sequences 5 times, increasing the number of random sequences increases the accuracy but increase the computational time and the memory load. Based on several tests, we considered that approximate estimate are worth trying for large dataset above  5000 / 10000 sequences, because the parallel computing start to subtantially decrease the computational time for minimal differences in outlier sequences detection.
}
\examples{
# Load the alignment file (class "alignment"), this is the first
# object of the list called Example_16S_outlier.
data(Example_16S_outlier)
Example_16S_outlier_align = Example_16S_outlier[[1]]
\dontrun{

# Running the function with an exact computation approach performing all the 
pairwise genetic comparisons among sequences. We used "raw" model to compute 
the genetic distances and considered a median distance as summary statistic 
to define the outlier threshold. We used a classic 1.5 interquartile range 
distance threshold to consider the outlier sequences.
S16_MisAlign = Clean.Seq(inputal = Example_16S_outlier_align, Outlier.Th = 1.5, 
output = "Example_16S_outlier_align.rm.out.", Evol.Model = "raw", Sum.Stat = "median", 
Iter.optim = FALSE, Re.Align = TRUE, nbthreads = 1, Approxim = FALSE, nb.seq = 25, 
nb.rep = 5) 
S16_MisAlign[[1]]  # 36 sequences are considered as outlier sequences the model detect
# To remove the file created while running the example do the following:
file.remove(c( "Example_16S_outlier_align.rm.out..Clean.AllSeq.fas", 
"Example_16S_outlier_align.rm.out..Clean.haplo.fas"))

Running the function with an approximate computation approach performing all 
the pairwise genetic comparisons among sequences. We used "raw" model to compute 
the genetic distances and considered a median distance as summary statistic to 
define the outlier threshold. We used a classic 1.5 interquartile range distance 
threshold to consider the outlier sequences. For the approximate comparison we used 
25 sequences randomly repeated 5 times as suggested by default.
S16_MisAlign.approx = Clean.Seq(inputal = Example_16S_outlier_align, Outlier.Th = 1.5, 
output = "Example_16S_outlier_align.rm.out.approx", Evol.Model = "raw", Sum.Stat = "median", 
Iter.optim = FALSE, Re.Align = TRUE, nbthreads = 5, Approxim = TRUE, nb.seq = 25, nb.rep = 5) 
S16_MisAlign.approx[[1]]  # 36 sequences are considered as outlier sequences the model detect
setdiff(S16_MisAlign.approx[[1]], S16_MisAlign[[1]]) # excat same finding considering the exact 
or the approximate search of outlier sequences.
To remove the file created while running the example do the following:
file.remove(c( "Example_16S_outlier_align.rm.out.approx.Clean.AllSeq.fas", 
"Example_16S_outlier_align.rm.out.approx.Clean.haplo.fas"))

}

}
