% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{detect.out.seq}
\alias{detect.out.seq}
\title{This function detects outlier sequences and is used internally by teh Clean.Seq function.}
\usage{
detect.out.seq(
  input = NULL,
  Evol.Model = Evol.Model,
  Sum.Stat = "median",
  Outlier.Th = 1.5,
  Approxim = FALSE,
  nb.seq = 25,
  nb.rep = 5,
  nbthreads = 1
)
}
\description{
This function helps to detect sequences that are abnormally distance form all other sequences in alignment. See the Clean.Seq function for a more detailled description.
}
\details{
For debug
rm(cytb2Sa, NewAlign.Noindels, NewAlign.Noindels2, Res1, cytb2.uniq, Res.optim, Seq.name.origin, cytb2, cytb2a, cytb2.toExportOutlier.Seqs)
}
