% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect.Outlier.Seq.R
\name{Detect.Outlier.Seq}
\alias{Detect.Outlier.Seq}
\title{Help to detect outlier sequences in an alignment}
\usage{
Detect.Outlier.Seq(
  inputal = NULL,
  Strat.DistMat = NULL,
  Dist.Th = NULL,
  output = NULL,
  Second.Outlier = NULL,
  Bitsc.Th = NULL
)
}
\arguments{
\item{inputal}{an object of class "alignment" (seqinr R package),
or "DNAbin" (ape R package) or the name (including the path if necessary) of the input fasta file.}

\item{Strat.DistMat}{can be 'MisAli', 'DivSeq' or 'Comb', see details.}

\item{Dist.Th}{distance threshold to detect the primary set of
outlier sequences (between [0-1]).}

\item{output}{name of the output table exported in the working directory.}

\item{Second.Outlier}{if 'TRUE', blasts the primary outlier sequences
against the local database of retrieved and aligned sequences
(i.e. the alignment provided to 'inputal') in order to detect secondary outliers.}

\item{Bitsc.Th}{bitscore similarity threshold to detect 'secondary'
outlier sequences, should be a value between [0-1].}
}
\value{
The function return a table for all the retained sequences with the following headers:
\itemize{
\item 'Query_SeqName' Name of the query sequence,
\item 'SeqLen_Query' length of the query sequence,
\item 'Hit_SeqName' name of the hit sequence,
\item 'Hit_SeqLen' length of the hit sequence,
\item 'evalue' E-value,
\item 'pident' means percentage of identical matches,
\item 'length' means alignment length,
\item 'mismatch' means number of mismatches,
\item 'gapopen' means number of gap openings,
\item 'qstart' means start of alignment in query,
\item 'qend' means end of alignment in query,
\item 'sstart' means start of alignment in subject,
\item 'send' means end of alignment in subject,
\item 'bitscore' means bitscore,
\item 'qcovs' means query coverage per subject,
}
(See \url{https://www.ncbi.nlm.nih.gov/books/NBK279675/} for more details and
options)
}
\description{
This function helps to detect some potential outlier sequences
(i.e. mis-aligned sequences, which might be caused by different problems such
as gene or species annotation problems, presence of paralogous sequences, etc)
that should be removed from the pool of sequences (per species and gene regions)
before selecting the best sequence.  This function is only provided as a tool
to HELP detect potential outlier sequences, great care must be taken to
ensure sequences are not erroneously omitted as potential outliers,
or remain included undetected outlier sequences. We recommend checking
alignments by eye to ensure your confidence in the retained sequences.
}
\details{
This function uses the following general strategy to detect outlier
sequences.  First, a distance matrix is computed among all sequences (see
different strategies available below), second a BIONJs tree (i.e. an improved
Neighbor joining method that allows missing values, Criscuolo & Gascuel 2008)
is built based on the distance matrix, and the tip to root distance is computed.
All the sequences above a certain distance threshold (we
use tip to root distances at least 50-60\%, see Chen et al. 2015 for similar thresholds)
are extracted as potential 'primary' outlier sequences.  In
addition, we can blast these primary outlier sequences back against the original alignment
converted into a local blast database (the alignement is automatically converted
by the function into a local database) to look for secondary outlier sequences that might
be similar to the primary sequences but just below the tip to root distance threshold.
All the sequences above a bitscore threshold (we use 80\%) are considered as potential 'secondary'
outliers. A bitscore threshold of 100\% corresponds to blasting the query sequence against itself.
We used the bitscore because it is not influenced by the size of the local database contrary to the E.value
(\url{https://www.ncbi.nlm.nih.gov/BLAST/tutorial/Altschul-1.html}).

Three options are available to compute the distance matrix among sequences.
The first option 'MisAli' uses the number of indelblocks to compute the
distance matrix in order to preferentially detect misaligned sequences
generating gaps in the alignment.  The second option 'DivSeq' uses the TN93
substitution model to compute the distance matrix in order to detect very
divergent sequences. The third option 'Comb' combines the two
distance matrices described above to detect both misaligned and very divergent sequence.
To do this, the two distance matrices are first re-scaled as percentages of the their highest
distances respectively and then summed. For a additional details see Eme et al. (2019) Appendix 4.

This function requires BLAST+ installed and in the PATH in order to detect 'secondary outliers'.
To download and install BLAST+ software locally go
to \url{https://blast.ncbi.nlm.nih.gov/Blast.cgi?CMD=Web&PAGE_TYPE=BlastDocs&DOC_TYPE=Download},
specific installation instructions for the different operating systems (OS) can be find
at \url{https://www.ncbi.nlm.nih.gov/books/NBK279671/}.
}
\examples{
# Load the alignment file (class "alignment"), this is the first
# object of the list called Example_16S_outlier.
data(Example_16S_outlier)
Example_16S_outlier_align = Example_16S_outlier[[1]]
\dontrun{

# Running the function with the 'Comb' option,
# a distance threshold of 0.6, and disabling the search for secondary outlier sequences.
S16_MisAlign0.6_1 = Detect.Outlier.Seq(inputal = Example_16S_outlier_align,
Strat.DistMat = "Comb", Dist.Th = 0.6, output = "Example_16S_outliers_1.txt",
Second.Outlier = "No")
dim(S16_MisAlign0.6_1) ### 36 sequences detected as primary outlier sequences.
head(S16_MisAlign0.6_1)

# The output table is also present as an external data table provided in the regPhylo r package
and can be accessed by the following code:
# a = system.file("extdata/ExampleOutliers/Example_16S_outliers_1.txt", package = "regPhylo")
# Example_16S_outliers_1 = read.delim(a, sep="\t", header = TRUE)


# Running the function with the 'Comb' option, a distance threshold of 0.6,
# and allowing a search for secondary outlier sequences using local blast database.
S16_MisAlign0.6_2 = Detect.Outlier.Seq(inputal = Example_16S_outlier_align,
Strat.DistMat = "Comb", Dist.Th = 0.6, output = "Example_16S_outliers_2.txt",
Second.Outlier = "Yes", Bitsc.Th = 0.8)
length(unique(S16_MisAlign0.6_2[,3])) ### 38 primary and secondary outlier sequences detected.
head(S16_MisAlign0.6_2)

# The output table is also present as an external data table provided in the regPhylo r package
and can be accessed by the following code:
# a = system.file("extdata/ExampleOutliers/Example_16S_outliers_2.txt", package = "regPhylo")
# Example_16S_outliers_2 = read.delim(a, sep="\t", header = TRUE)

# To remove the file created while running the example do the following:
file.remove(c( "Example_16S_outliers_1.txt", "Example_16S_outliers_2.txt"))

}

}
\references{
Criscuolo & Gascuel 2008, DOI: 10.1186/1471-21.5-9-166

Chen et al. 2015, DOI: 10.1093/sysbio/syv059

Eme et al. (2019). An integrated pathway for building regional
phylogenies for ecological studies. \emph{Global Ecology and Biogeography}, Accepted.
}
