% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{DNAbin2alignment}
\alias{DNAbin2alignment}
\title{Convert a DNAbin object (ape format) to an alignment object (seqinr format)}
\usage{
DNAbin2alignment(input)
}
\arguments{
\item{input}{a DNAbin object}
}
\description{
This function converts a DNAbin object (ape format) loaded with the ape::read.FASTA 
to an alignment object (seqinr format). It avoids the conversion from character 
to numeric in the DNAbin object.
}
