% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First.Align.All.R
\name{First.Align.All}
\alias{First.Align.All}
\title{Reverse complement and align the sequences using MAFFT and PASTA}
\usage{
First.Align.All(
  input = NULL,
  output = NULL,
  nthread = NULL,
  nthread.per.align = NULL,
  methods = NULL,
  Mafft.path = NULL
)
}
\arguments{
\item{input}{path to the folder storing the files of interest. The files must have
the extension '.fas' and the name of the gene region must be immediately before
the extension separated from the rest of the name by '_' (e.g.
'Alignment_co1.fas').}

\item{output}{path to the folder storing the final alignments,
(the output folder is created directly by the function). The name of the
program appears as a prefix in the name of the alignment file using '_' as a
separator. The function also exports a table into the input folder with the name
of all the sequences per gene region that have been reverse complemented.}

\item{nthread}{number of threads used to run the alignment software in parallel for
the different gene regions.}

\item{nthread.per.align}{number of threads used to run the alignment software for a 
single alignment at the time. This option is NULL by default but is specially suited 
when performing an alignment for a single DNA marker with many DNA sequences.}

\item{methods}{programs used to align the sequences, the function used Mafft-FFTNS1
to reverse complement and align quickly the sequences, but other programs can be also
used simultaneously, such as  Mafft-fftnsi and PASTA, c("mafftfftnsi", "pasta", 
"mafftfftns2", "mafft_auto"). The option mafft_auto choose automatically the best 
algorithm among fftns1, fftns2, fftnsi or l-nsi.}

\item{Mafft.path}{for the Windows platform, a character string which provides the path
to the mafft executable
(e.g. Mafft.path = "C:/Users/deme/Documents/Programs/Mafft/mafft-7.409-win64-signed/mafft-win/mafft")
(see examples below). Eventhough Mafft is not the alignment you
are after, mafft is necessary to reverse complement the sequences and the Mafft.path must
be completed in all cases. For Linux the mafft software must be in the $PATH.}
}
\value{
The function returns the alignment in the output folder using the
name of the program as a prefix of the alignment (i.e. "mafftfftnsi_Alig_co1.fas").
The function also returns a table called "ListSeq_RevCompl_FirstAlignAll.txt"
into the output folder listing the sequence names that have been reverse
complemented by Mafft, the same table is exported into the R environment as well.
}
\description{
First, this function aligns all the sequences for each distinct gene region
using Mafft v7.222 fftns1 (Katoh et al. 2005) with the
option \emph{--adjustdirection} (i.e. to automatically detect the direction of the sequences
and reverse complement the sequences if necessary). The function adds a '_R_' as a
prefix to the sequence name if it is reverse complemented. Second, the function runs Mafft fftnsi (Katoh et
al. 2005), and PASTA (Mirarab et al. 2015) in parallel and alphabetically orders
the sequences within each alignment (i.e. for each gene region).
\strong{Note:} PASTA is not available to run for Windows users.
}
\details{
PASTA (as Sate I and II) may be better at handling very
divergent sequences and at dealing with saturated phylogenetic signal because
it splits the alignment into small chunks of similar sequences and performs
profile alignments among those smaller chunk of sequences (Roquet et al. 2013).

The 'output', 'input' and 'nthread' or 'nthread.per.align' objects need
to be present in the R environment before running the function because the
function runs in parallel using the R package 'parallel', see example below the
function.

The function requires that Mafft and PASTA are installed and in the PATH.
\itemize{ \item For Mafft, download for the different OS and all documentation are available at
\url{https://mafft.cbrc.jp/alignment/software/}, (for Windows OS we have only tested the
"All-in-one package for Windows" version of Maftt).
\item For PASTA, download and instructions for installation are provided on the Github page
available at: \url{https://github.com/smirarab/pasta}.
}
}
\examples{
# The input, output and nthread object have to be present in
# the R global environment before running the function.
\dontrun{
# To demonstrate the function we provided small example alignments in the
# inst/extdata/FirstToAlign folder available with the package.
# These alignments are exported from the Seq.DF5 dataset, for
# the co1 and 16s regions. Notice that the 16s has only one
# sequence and this is explained in the warning message.

# To run the example, copy the input alignment files
# provided by the package to a temporary directory in your
# current working directory.
src.dir = system.file("extdata/FirstToAlign", package = "regPhylo")
dir.create("TempDir.FirstToAlign")
# Set up the path to the TempDir folder.
dest.dir = paste(getwd(), "/TempDir.FirstToAlign", sep="")
file.names <- dir(src.dir)
# Copy all the files stored in regPhylo/extdata/FirstToAlign"
# into a temporary folder.
sapply(file.names, function(x) {
file.copy(from = paste(src.dir, x, sep = "/"),
to = paste(dest.dir, x, sep = "/"),
overwrite = FALSE) })

# Remove the empty file created for the folder name when
# importing the data file
file.remove("TempDir.FirstToAlign/FirstAligned")

input = "TempDir.FirstToAlign"
output = "TempDir.FirstToAlign/FirstAligned"
nthread = 2
methods = c("mafftfftnsi", "pasta")

First.Align.All(input = input, output =
"TempDir.FirstToAlign/FirstAligned", nthread = 2,
methods = c("mafftfftnsi", "pasta"))


#### To run FirstAlign.All on Windows OS. ###

input = "TempDir.FirstToAlign"
output = "TempDir.FirstToAlign/FirstAligned"
nthread = 2
methods = "mafftfftnsi"
First.Align.All(input = input, output =
"TempDir.FirstToAlign/FirstAligned", nthread = 2,
methods = "mafftfftnsi",
Mafft.path = "C:/Users/deme/Documents/Programs/Mafft/mafft-7.409-win64-signed/mafft-win/mafft")

#### end specifications to run FirstAlign.All on Windows OS. ####


# To remove the file created while running the example do the following:
unlink("TempDir.FirstToAlign", recursive = TRUE)

}
}
\references{
Katoh et al. 2005, DOI: 10.1093/nar/gki198

Mirarab et al. 2015 DOI: 10.1089/cmb.1014.0156

Roquet et al. 2013 DOI: 10.1111/j.1600-0587.2012.07773.x
}
