% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoCodeName.R
\name{GeoCodeName}
\alias{GeoCodeName}
\title{Retrieve geographic coordinates from place names using nominatim openstreetmap API}
\usage{
GeoCodeName(input = NULL, output = NULL)
}
\arguments{
\item{input}{a table coming from the function GeoCoord.WGS84() or from
Query.GeOMe.XY.R()}

\item{output}{name of the output table exported in the working directory.}
}
\value{
Two identical tables are exported, one in the R environment and one in the working directory.
These tables includes additional fields: Location_used= name
of the location used to infer the geographic coordinates,
Geo_accuracy= precision of the coordinates, with
'Inferred'= imprecised geographic coordinates inferred from the Location name,
'From_DB'= accurate geographic coordinates (provided by the databases), and
'NoLocationFound'= the geographic coordinates cannot be inferred from the place
name.
}
\description{
This function uses the nominatim openstreetmap API
(https://nominatim.openstreetmap.org) to retrieve
geographic coordinates for the sequences without geographic coordinates but
carrying information about their sampling location
}
\details{
Since the RGoogleMap API required a paying key (changed policy the July 16 2018)
we used the nominatim openstreetmap API.
}
\examples{
## Import the example data coming from GeoCoord.WGS84 function.
\dontrun{
data(Seq.Diastocapen)
Seq.DF1 = Seq.Diastocapen$Seq.DF1

# Run the function to retrieve geographic coordinates for sequences without coordinates.
Seq.DF3=GeoCodeName(input = Seq.DF1, output = "Seq.DF3.txt")
}

}
