% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.taxo.classif.R
\name{Get.classif}
\alias{Get.classif}
\title{Arrange the classification information exracted from the itis, gbif and bold.}
\usage{
Get.classif(input = NULL, Tax.list = NULL)
}
\arguments{
\item{input}{is the output of the /emph{taxize::classification} function, which is 
a list of data.frames with the taxonomic classifiation of the supplied taxa.}

\item{Tax.list}{is a data frame with two columns the first include the accepted taxonomic name 
and the second the synonyms taxonomic names.}
}
\description{
This function arranges the classification from the ncbi or itis database using the taxize R package. It keeps the class, order, family, genus and species names with their associated txid. This function work internally for the \emph{Get.taxo.classif} function.
}
