% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.taxo.classif.R
\name{get.db.ids}
\alias{get.db.ids}
\title{From a taxonomic name get the id of taxon from different open-acces databases, ncbid, gbif, bold, itis, and taxref}
\usage{
get.db.ids(
  taxon.name = NULL,
  db = NULL,
  ncbi.api.key = NULL,
  input.ID.Rank = "Genus",
  taxref.fuzzymatch = TRUE
)
}
\arguments{
\item{taxon.name}{= the name of the taxon of interest eg. "Drusus"}

\item{db}{=  the names of the taxonomic data base, among "itis", "ncbi", "tropicos", "gbif", "eol", "nbn", "pow", "bold", and taxref (via rtaxref Rpackage to get acess to the taxref from the French INPN database)).}

\item{ncbi.api.key}{a string with the ncbi api key (personal to the user, it should be created via NCBI website)}

\item{input.ID.Rank}{the Rank of the ID requested eg. "Genus" or "Species"}

\item{taxref.fuzzymatch}{if TRUE (default), a fuzzy search is also performed for the taxref database if the 
request using an exact match is unsuccessful.}
}
\value{
the function returns a vector of the taxonomic id of the corresponding databases.
}
\description{
This function from a taxonomic name (eg. "Plecopetra") gets the id of a taxon from
different open acess data base, ncbid, gbif, bold, itis, and taxref (this function is used
internally by Get.taxo.classif.Multi
}
