% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.taxo.classif.R
\name{Get.taxo.classif.Multi}
\alias{Get.taxo.classif.Multi}
\title{Extract a taxonomic information and check for synonyms in the NCBI, itis, gbif, bold and
taxref (MNHN) databases}
\usage{
Get.taxo.classif.Multi(
  Taxa.list = NULL,
  dbs = NULL,
  api_keys = NULL,
  ncbi.api.key = NULL,
  ID.Rank = "Genus",
  Save.Intermediate.ID.rank = NULL
)
}
\arguments{
\item{ID.Rank}{the name of the Rank of the Taxa.list, it can be either "Species", or any other
levels (by default, the function considers the "Genus"), however in the results only the Species,
Genus, Family, Order and Class hierarchical levels are reported.}

\item{Save.Intermediate.ID.rank}{the path of the folder to store the results of the different
Taxa.list.}

\item{input.id}{the taxid of the taxa of interest, (extract all the descendant species, from the
ncbi, itis, gbif, bold and taxref database.}

\item{db}{the database to query either "ncbi", "itis", "gbif", "bold", or "taxref" (TAXREFV1.5,
from the rtaxref R package).}

\item{downto}{hierarchical level to look for descent, by default it is species. see function
downstream from the taxize R package.}

\item{local.db}{THIS OPTION IS CURRENTLY INVALID, WORK IN PROGRESS, if TRUE (by default FALSE),it use local imported database to query the
classification and performed request, can be much faster for important request. 
This is working with "ncbi","gbif" (not for the search of synonyms). 
If FALSE the function use the API to query the online server.}

\item{input.ID.Rank}{the Rank of the ID requested}
}
\value{
The function returns a table with the following fields
c("species.valid", "species.syn", "genus", "family", "order", "class", "txid.species.valid",
"txid.species.syn", "txid.genus", "txid.family", "txid.order", "txid.class", "database",
"extraction.date").
}
\description{
This functions extracts the species list, the classification and the synonyms 
species name of the accepted species names from the ncbi taxonomy, itis, gbif, bold, taxref (MNHN
INPN) databases from a taxid of a taxon from a hierarchical order from species level to higher
taxonomic levels (eg. "Genus", "Family" , "Order"..), obtain by the get_ids("plecoptera", db =
"ncbi") function for instance.
}
\details{
This function allows to check for the presence of the synonyms found in a database, their
are also cross checked in the other databases. This function allow to loop the search over 
multiple species or genera or family at the same time over the different databases.
}
