% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.taxo.classif.R
\name{get_classif.ncbi}
\alias{get_classif.ncbi}
\title{Get Classification table and synonyms from NCBI taxonomy database}
\usage{
get_classif.ncbi(input = NULL, api_key = NULL)
}
\arguments{
\item{input}{a vector with the ncbi taxonomic ID of the taxa of interest.}

\item{api_key}{a vector with the personnal api key to get a better acces to the NCBI API.}
}
\value{
The function returns a table with the following fields
c("species.valid", "species.syn", "genus", "family", "order", "class", "txid.species.valid",
"txid.species.syn", "txid.genus", "txid.family", "txid.order", "txid.class", "database",
"extraction.date").
}
\description{
This function gets the classification and the synonyms taxonomic name of the taxa of
interest from the NCBI taxonomy database using the rentrez R package. This function is used
internally by \emph{Get.taxo.classif} and \emph{Get.taxo.classif.Multi} functions.
}
\details{
This function works only if the api_key of the user is provided in the option
\emph{api_key}.
}
