% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSeq_BOLD.R
\name{GetSeq_BOLD.local.DT}
\alias{GetSeq_BOLD.local.DT}
\title{Extract DNA sequences and metadata from a local download of BOLD (Barcode Of Life Database)}
\usage{
GetSeq_BOLD.local.DT(splist = NULL, filename = NULL, Path.BOLD.tsv = NULL)
}
\arguments{
\item{splist}{a two column table. The first column contains all the binomial species names including
synonyms and the second column contains only the names that will be reported
in the output table as 'species_name'.}

\item{filename}{name of the output table. The format of the output follows the
format provided by the bold R package.}

\item{Path.BOLD.tsv}{name of the data.frame (R object) containing the Barcode Of Life Database BOLD, or 
the path of the tsv file of the BOLD that can be downloaded at https://boldsystems.org/data/data-packages/}
}
\value{
The function returns three tables: a first table, indicated by "filename" exported in a folder
indicated within the "filename", it contains the raw output from bold package plus an additional column with the date of extraction
The function returns into the R environment a list of two tables that are exported including a summary table reporting
the number of occurrences retrieved for each species, and a table reporting all
the data retrieved from bold after removing record without DNA sequence.
}
\description{
This function extracts sequences and associated information from a local version of the Barcode Of
Life Database (BOLD) for all the species in the species list
}
