% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSeq_BOLD.R
\name{GetSeq_BOLD}
\alias{GetSeq_BOLD}
\title{Extract DNA sequences and metadata from BOLD (Barcode Of Life Database)}
\usage{
GetSeq_BOLD(splist = NULL, filename = NULL)
}
\arguments{
\item{splist}{a two column table, the first column contains all the binomial species name including
synonymes and the second column contains only the names that will be reported
in the output table as 'species_name'.}

\item{filename}{name of the output table. The format of the output follows the
format provided by the bold R package.}
}
\value{
The function returns 3 tables: a first table with the raw output from bold package plus
an additional column with the date of extraction is exported into the working directory
(this table reports only metadata when the DNA sequence is present), and a
list of two tables exported in the R environment, a summary table reporting the number of
occurrences retrieved for each species, and the second table reporting all
the data retrieved from bold (including metadata in absence of DNA sequence,
as such the data present in this table and the table exported in the working directory might differ)
}
\description{
This function extracts sequences and associated information from the Barcode Of
Life Database (BOLD) for all the species in the species list
}
\examples{
# a table with 1 species
Splist = cbind(TaxID = c("Diastobranchus capensis"), Species.Name = c("Diastobranchus capensis"))
\dontrun{
# Run the function considering extracting all DNA sequences and associated metadata
BOLD.output = GetSeq_BOLD(splist = Splist, filename = "output.BOLD.txt")

# The output can be loaded doing the following
data(Seq.Diastocapen)
BOLD.output = Seq.Diastocapen$Seq.BOLD
}
}
