% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.DNA.extraction.R
\name{GetSeqNumber_NCBI_taxid}
\alias{GetSeqNumber_NCBI_taxid}
\title{Get the number of DNA sequences available for taxa of interest in NCBI.}
\usage{
GetSeqNumber_NCBI_taxid(splist = NULL, gene = "ALL", chunk_size = 50)
}
\arguments{
\item{splist}{a two column table with the unique NCBI taxonomic ID (i.e. taxid) in
the first column and the accepted binomial species name in the second column
which will be retained as 'TaxaName' in the output file.}

\item{gene}{can be a particular gene region or 'ALL' for all entries
of that species. Please note that this will only retrieve genes
with the same name in Genbank. For example, "COX1" will not return the
same entries as "COI". Default "ALL"}

\item{chunk_size}{number of records to be downloaded at a time from Genbank.
If connection drops try decreasing this value to 20 or smaller. Default 50.}
}
\description{
This function provides the number of DNA sequences available for the taxa of interest
in th NCBI database, this function is used internally in the Pipeline.DNA.extraction function
}
