% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTaxnum_NCBI.R
\name{GetTaxnum_NCBI}
\alias{GetTaxnum_NCBI}
\title{Retrieves NCBI taxa ID}
\usage{
GetTaxnum_NCBI(splist = NULL)
}
\arguments{
\item{splist}{a character vector of species as "Genus species" to query the NCBI Taxonomy database.}
}
\value{
This function returns a table with the species names and the NCBI taxa ID
The output is retrieved from the following website: https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi
}
\description{
This function retrieves the NCBI taxa ID from the NCBI taxonomy database
given a list of "Gen sp" values
}
\details{
This function requires httr
}
\examples{
Create a species list and retrieve NCBI:txid codes


\dontrun{
Sp.List = c("Aplodactylus etheridgii", "Aplodactylus arctidens", "Asterorhombus filifer")
taxReport <- GetTaxnum_NCBI(splist = Sp.List)
write.table(taxReport, file="tax_report.txt", sep= "\t", quote = FALSE, row.names = FALSE)

}
}
