% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.Overlap.R
\name{Matrix.Overlap}
\alias{Matrix.Overlap}
\title{Estimate the species overlap between each per of selected gene region}
\usage{
Matrix.Overlap(input = NULL, gene.Sel = NULL)
}
\arguments{
\item{input}{the species-by-gene matrix which is provided as the first object in the}

\item{gene.Sel}{a vector of a selected gene region
(gene names have to be consistent with the header of the
table with the suffix '_CleanDataset.txt' exported by the function
SpeciesGeneMat_Bl.R).  Example (Cytochrome c oxydase subunit 1, should be
written 'co1' and not 'COI' or 'COX1').}
}
\value{
The function returns a first matrix providing the raw
number of species overlapping between pairs of gene regions, and a second
matrix that provides the proportion of species overlap between pairs of gene
regions, the proportion is computed as a percentage of the richest gene region
(i.e. the gene region that is represented in the greatest number of species).
}
\description{
This function estimates the number of species overlapping between each pair of
selected gene regions.
}
\examples{
# Load a species-by-gene matrix for instance exported by SpeciesGeneMat.Bl function.
data(Seq.DF4) ## the first object of the list is the Species-by-gene matrix

# Run the function.
Matrix.Overlap(input = Seq.DF4[[1]], gene.Sel = c("co1", "16srrna"))


}
