% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi.Align.R
\name{Multi.Align}
\alias{Multi.Align}
\title{Align sequences using multiple alignment programs}
\usage{
Multi.Align(input = NULL, output = NULL, nthread = NULL,
  methods = NULL)
}
\arguments{
\item{input}{path to the folder storing the files of interest for further alignment,
(the files should have the extension '.fas' and the name of the gene region
should be immediately before the extension separated from the rest of the name
by '_', e.g.  'Alignment_co1.fas').}

\item{output}{path to the folder storing the
final alignments (the output folder is created directly by the function). The
name of the program appears as a prefix in the name of the alignment file using
'_' as separator. The function also exports a table in the input folder with
the name of all the sequences per gene region that have been reverse
complemented.}

\item{nthread}{number of threads used to run the alignment software in}

\item{methods}{programs used to aligned the sequences, the function used Mafft-FFTNS1 anyway,
but other programs can be used simultaneously, such as Mafft-fftns2, Mafft-fftnsi,
Muscle, Prank and PASTA, c("mafftfftns2", "mafftfftnsi", "muscle", "prank", "pasta").}
}
\description{
This function runs several alignment softwares, Muscle v3.8.31 (Edgar 2004),
Mafft v7.222 fftns1, fftns2, fftnsi (Katoh et al. 2005), Prank v.140603 (Loytymoja & Goldman 2008)
and PASTA (Mirarab et al. 2015) in parallel.
}
\details{
First, for each distinct gene region, the function aligns all the sequences
using Mafft v7.222 fftns1 (Katoh et al. 2005)
with the option --adjustdirection to automatically detect the direction of the
sequences and reverse complement the sequences if necessary (the function adds
a '_R_' as prefix on the sequence name). Second, the function runs Mafft
fftns2, fftnsi, Prank and PASTA, and alphabetically orders the sequences within
each of all the alignments for the same gene.

The function requires that Muscle, Mafft, Prank and Pasta are installed and in
the PATH.

The 'output', 'input' and 'nthread' objects need to be present in
the R environment before running the function
because the function runs in parallel using the R package 'parallel', see
example.
}
\examples{
# Load in the R environment the object used by the function.
# Here the toy example consist of four alignment (co1, 12srrna, cytb, rag1)
# for a subset of 16 species.
\dontrun{

# To run the example it might be better to copy the input alignment files
# provided by the package to a temporary directory created into the
# current working directory.
src.dir = system.file("extdata/multi.align", package = "regPhylo")
dir.create("TempDir.Multi.aligned")
# Set up the path of the TempDir folder.
dest.dir = paste(getwd(), "/TempDir.Multi.aligned", sep="")
file.names <- dir(src.dir)
# Copy all the files stored in regPhylo/extdata/multi.align"
# into a temporary folder.
sapply(file.names, function(x) {
file.copy(from = paste(src.dir, x, sep = "/"),
to = paste(dest.dir, x, sep = "/"),
overwrite = FALSE) })

output = "multi.alignedTest"
nthread = 3
methods = c("mafftfftnsi", "pasta")
input = "TempDir.Multi.aligned"
# run the function using two additional alignmenet program "mafftfftnsi" and "pasta"
Multi.Align(input = input, output = "multi.alignedTest", nthread = 3,
methods = c("mafftfftnsi", "pasta"))
list.file("multi.alignedTest")

# To clean the file created while running the example do the following:
# Remove the temporary folder
unlink("TempDir.Multi.aligned", recursive = TRUE)
# Remove the folder with Gblocks outputs
unlink("multi.alignedTest", recursive = TRUE)


# The output files are also present as an external data in the regPhylo package
# and can be accessed by the following code:
# a = system.file("extdata/multi.align/multi.aligned", package = "regPhylo")
# list.files(a)
}

}
\references{
Edgar 2004, DOI: 10.1093/nar/gkh340

Katoh et al. 2005, DOI: 10.1093/nar/gki198

Loytymoja & Goldman 2008, DOI: 10.1126/science.1158395

Mirarab et al. 2015, DOI: 10.1089/cmb.1014.0156
}
