% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mumsa.Comp.R
\name{Mumsa.Comp}
\alias{Mumsa.Comp}
\title{Compare congruence among multiple alignments: a wrapper for MUMSA software}
\usage{
Mumsa.Comp(input = NULL, output = NULL, remove.empty.align = NULL)
}
\arguments{
\item{input}{path to the folder storing the alignments. The names of the alignments
must follow the following template: name of the alignment program, the name of
interest, and the name of the gene region just before the extension, and using
'_' as a separator.  The alignments have to be in fasta format (extension
'.fas').}

\item{output}{path and name of the file where you want to store the
results of MUMSA analysis. The output table reports the AOS for each gene
region and the MOS for each alignment and each gene region.}

\item{remove.empty.align}{If TRUE, the empty alignments are excluded from the computation.}
}
\value{
A tables with the MOS for each alignment and programs and AOS for each
alignment are exported as a txt file.
}
\description{
This function compares multiple alignments using the MUMSA software (Lassmann &
Sonnhammer 2005) and reports the MOS (Multiple
Overlap Score) and the AOS (Average Overlap Score) statistics.
\strong{Note:} This function is currently not available on Windows platform (Windows user can use
the online web server available at: \url{http://msa.cgb.ki.se/cgi-bin/msa.cgi}).
}
\details{
The MOS is provided for each alignment enabling us to check the support (more exactly the
consensuality) of those alignments. The alignment with the highest MOS is
supposed to be the most consensual alignment (Lassmann & Sonnhammer 2005).
Alignments with equal MOS are equally supported. The AOS gives an idea as to
whether the sequences are too divergent to be aligned. A AOS less than 0.5
indicates that the sequences are too divergent to be properly aligned
(potentially caused by saturation effects).

The function requires that the MUMSA software is installed and in the PATH.
Download from \url{http://msa.cgb.ki.se/cgi-bin/msa.cgi}, paste the archive into the desired
folder, extract the archive and then compile the program using the make commands.
}
\examples{
# Run the function without removing empty alignments
# Here the example consists of the alignment of four gene regions
# (co1, 12srrna, cytb, rag1) for 16 species.
# Each gene region was aligned using three programs mafftfftns1,
# mafftfftnsi, and pasta.
\dontrun{
input = system.file("extdata/multi.align/multi.aligned", package = "regPhylo")
Mumsa.Comp(input = input, output = "Mumsa_output",
remove.empty.align = FALSE)
# The function report an error message if some alignment are empty.
# Run the function removing empty alignments
input = system.file("extdata/multi.align/multi.aligned", package = "regPhylo")
Mumsa.Comp(input = input, output = "Mumsa_output",
remove.empty.align = TRUE)

# The output file is also present as an external data table in the regPhylo package
# and can be accessed by the following code:
# a = system.file("extdata", "Mumsa_output.txt", package = "regPhylo")
# a = read.delim(a, sep="\t", header = TRUE)

# To remove the file created while running the example do the following:
file.remove("Mumsa_output.txt")

}

}
\references{
Lassmann & Sonnhammer 2005, DOI: 10.1093/nar/gki1020
}
