% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.DNA.extraction.R
\name{Pipeline.DNA.extraction}
\alias{Pipeline.DNA.extraction}
\title{A pipeline to automat the retriveal and organisation of DNA sequences and metadata 
from NCBI and BOLD databases from a list fo taxa of interest.}
\usage{
Pipeline.DNA.extraction(
  input.ncbi = NULL,
  input.bold = NULL,
  gene = "ALL",
  max.seq = 10000,
  Path.output = NULL,
  Nb.DNA.marker = 1,
  Steps = c(1, 2),
  nthread = 1,
  methods = "mafftfftns2",
  nthread.per.align = NULL
)
}
\arguments{
\item{input.ncbi}{a two column table with the unique NCBI taxonomic ID (i.e. taxid) in
the first column (incluidng the taxonomic ID of synonyms species names) 
and the accepted binomial species name in the second column
which will be retained as 'TaxaName' in the output file.}

\item{input.bold}{a two column table with the first column containing all 
the binomial species names including synonyms and the second column contains 
only the names that will be reported in the output table as 'species_name'.}

\item{gene}{can be a particular gene region or 'ALL' for all entries
of that species. Please note that this will only retrieve genes 
with the same name in Genbank. For example, "COX1" will not return the 
same entries as "COI". Default "ALL"}

\item{Path.output}{the path to the output folder within which all the intermediate and final
results files will be stored.}

\item{Nb.DNA.marker}{can be a numeric (eg. 4) to select the first gene with the highest species
coverage, it can be a gene name (eg. "co1") or it an be "Min.All.Sp.cov" in order to include the
minimum gene region to get 100% species coverage.}

\item{Steps}{is a vector of numeric value from 1 to 3, with 1 extract the DNA squences and 
associated metadata from NCBI and BOLD, dereplicate the data, improve the metada location, and
build the Species by Gene matrix, step 2 selects the gene of interest and exports the metadata 
and the alignments in fasta format, step 3 performs a first alignment approach of the selected gene
of interest to detect sequence entered in wrong direction and perfomed automatic reverse
complementation. The steps are perfomed sequentially so the steps 2 cannot be performed alone
without perfoming the steps 1.}

\item{nthread, }{if steps 3 is present in the option Step (eg. Steps = c(1,2,3)) then we can choose #' how many thread we wouldlike to use to perform the first alignment steps in case there are multiple
markers to align.}

\item{methods}{if steps 3 is present in the option Step (eg. Steps = c(1,2,3)) then we can choose
the method used for the alignment approach, it can be "mafft_auto", "mafftfftnsi", "mafftfftns2",
and "pasta". See the function "First.Align" from the regPhylo R package. It can also be a
combination of several methods eg. c("mafftfftnsi", "mafftfftns2", "pasta").}

\item{nthread.per.align}{if steps 3 is present in the option Step (eg. Steps = c(1,2,3)) then we
can choose how many threads we would like to use to perform the first alignment step. This option
is interesting when few alignments have many sequence > 1000 DNA sequences.}
}
\value{
this function returns a set of intermediate and final files and folders within the folder
precised in the Path.output option. The structure of the folders is organised as follow, 
The function return a list of 4 objects using the following headers:
\itemize{
\item the "Raw.Data" folder containing the raw data files of NCBI ("Seq.NCBI.txt") and BOLD
(Seq.BOLD.txt") DNA and metadata extraction, the combine file of DNA sequences from NCBI an BOLD
after dereplication ("Seq_NCBI_BOLD.txt"), the intermediate files with all the improve
georeferencing ("Seq_NCBI_BOLD_Geo1.txt", "Seq_NCBI_BOLD_Geo2.txt"), the files of the 
"species-"by-gene" matrix ("Matrix.Sp.DNA.NaiveSpDNA_Mat.txt") including files with all the
metadata ("Matrix.Sp.DNA.Naive_CleanDataset.txt"), and summary files of the species-by-gene matrix
from a DNA perspective ("Matrix.Sp.DNA.NaiveSummary_DNApers.txt") and from a species perspective ("Matrix.Sp.DNA.NaiveSummary_SPECIESpers.txt").
\item the "Selected.DNA" folder containing the non-aligned fasta files of the selected DNA regions
and the metadata of all the selected DNA sequences ("Align_InfoTabSelSeq.txt"), and the folder 
"FirstAlign"if the step 3 of the pipeline was performed. This folder contains the alignments of 
the selected DNA region from the different alignment software specify in the option "methods".
\item the folder "Selected.META" contains a file of the Metadata of the sequences of the 
selected DNA regions ("MetaData.Select.DNA.Select.DNA.txt").
}
}
\description{
This function allows to extract DNA sequences and metadata from NCBI and BOLD,
dereplicates the data from both databases, improve the georefercing of DNA sequences 
using openstreetmap API, buit a species by gene matrix, select the DNA region with the best
taxonmic coverage, export DNA sequences and the assocated metadata for selected Gene region, 
select the best DNA seqeunces per taxa and perform first alignment allowing automatic detection 
of sequences entered in a wrong direction.
}
\details{
this function will require a specific NCBI personal API key in order to 
get a faster access to the NCBI server.

This function build a pipeline of the following function "GetSeqNumber_NCBI_taxid" (to detect the number of sequence per taxon and identify the model species with too many 
sequence to download), "GetSeqInfo_NCBI_taxid" (get the DNA and the metadata from NCBI),
"GetSeq_BOLD" (get the DNA and the metadata from BOLD), "Congr.NCBI.BOLD.perReposit" (dereplicate
and improve metadata from NCBI and BOLD), "GeoCoord.WGS84" (homogeneize the geographic coordinates),
"GeoCodeName" (improve the georeferencing ofthe data with just locality names using the
openstreetmap API), "SpeciesGeneMat.Bl" (build the species-by-gene matrix),
"SelGene.MaxSpCov" (identify the DNA region maximiasing the sepcies coverage), "Select.DNA" (expot
the metadata of the selected DNA region, including the extraction of the DNA region within larger
gene region such as the COI within the mitogenome), "SelBestSeq" (select the DNA sequences based on
several criteria including some geographic proximity to some geographic region of interest, and
export the alignments in fasta format), "First.Align.All" (detection of sequences that are entered
in the wrong direction and first alignments using Mafft and pasta software).
When a taxa is detected with too many sequences ste by the max.seq option (10 000sequences by
default), these taxa are momentarily excluded and the species-by-gene matrix is performed without
them to identify the DNA region with the higeth taxonomic coverage, then once those region
identified, the DNA sequences and metadat of those region are extracted for the taxa initially
recognised with too many sequences. This function allows to run different steps, see the definition of the "Steps" parameter.
}
