% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.GeOMe.XY.R
\name{Query.GeOMe.XY.R}
\alias{Query.GeOMe.XY.R}
\title{Retrieve accurrate geographic coordinates from GeOMe database.}
\usage{
Query.GeOMe.XY.R(input = NULL, Phylum = NULL, output = NULL)
}
\arguments{
\item{input}{is a table coming from the function \code{\link{GeoCoord.WGS84}}.}

\item{Phylum}{is the name of the phylum}

\item{output}{name of the output table exported in the working directory.}
}
\value{
Two identical tables are exported; one in the R environment
and one in the working directory with the output name.
The output is similar to the input table except that the fields
Lat Long have been updated with proper geographic coordinates if they are
 found in GeOMe, finally the field 'OriginDatabase' is amended by
'_XY-GeOMe' for the relevant sequences.
}
\description{
This function queries the GeOMe database to retrieve location
metadata (i.e. geographic coordinates) for sequences lacking geographic coordinates
(\url{https://www.geome-db.org/} (Deck et al. 2017).
}
\details{
The names of the accepted phyla names in GeOMe are available
at the following web page: (\url{https://geome-db.org/query}).
In this query page, see the field 'Phylum' to see the phyla
available in GeOMe.
}
\examples{
# Import Seq.DF1
#' data(Seq.Diastocapen)
Seq.DF1 = Seq.Diastocapen$Seq.DF1
\dontrun{
# Run the function
Seq.DF2 = Query.GeOMe.XY.R(input = Seq.DF1, Phylum = "Chordata", output = "Seq.DF2.txt")
}

}
\references{
Deck et al. 2017, DOI: 10.1371/journal.pbio.2002925
}
