% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{rm.del.gap}
\alias{rm.del.gap}
\title{Function to remove sequences (or loci) with rare insertion, or containing too many ambiguous nucleotide, or sequences that are too short.}
\usage{
rm.del.gap(
  input = NULL,
  Remove.del.gaps = TRUE,
  Remove.Seq.Indels = FALSE,
  Nb.Seq.Indels = 2,
  Remove.Indels.Only = FALSE,
  Remove.Short.Seq = FALSE,
  Max.Per.missing = 50,
  Remove.Seq.TooManyAmbig = FALSE,
  Percent.Ambig = 30,
  Remove.Loc.Low.Freq = FALSE,
  Minimal.Locus.Freq = 30
)
}
\description{
This function helps to detect and remove sequences and loci with rare insertions, or containing too many ambiguous nucleotides, or sequences that are too short.
}
