% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.taxo.classif.R
\name{rt_taxa_classification}
\alias{rt_taxa_classification}
\title{Extract the classification of taxa from the TAXREF via the API.}
\usage{
rt_taxa_classification(id)
}
\arguments{
\item{id}{is a numeric value of the TAXREF id of the taxa.}
}
\description{
This function extracts the classification of a taxa from the TAXREF via the API.
t=This function is derived from the rtaxref R package https://github.com/Rekyt/rtaxref
}
