% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelGene.MaxSpCov.R
\name{SelGene.MaxSpCov}
\alias{SelGene.MaxSpCov}
\title{Estimate the minimal number of gene region to maximise species coverage}
\usage{
SelGene.MaxSpCov(input = NULL, NBGene = NULL)
}
\arguments{
\item{input}{the species-by-gene matrix which is provided as the first object in the}

\item{NBGene}{either a number of genes that we would like to include optimizing
the species coverage, or a vector of pre-defined gene regions
or NULL when the function is turned off.}
}
\description{
This function returns additional information including the number of genes
required to have 100% species coverage, the list of those genes, and a species
accumulation curve from a gene perspective giving an idea of the contribution
of each gene to maximise the species coverage.
}
\details{
For a pre-selected number of
genes, the function returns a list of the gene names maximizing the species
coverage and the species coverage. For a preselected set of genes the function
provides the species coverage, the list of species potentially missing, and the
minimum list of genes enabling us to have 100% coverage.
}
\examples{
# Load a Species-by-gene matrix exported from the SpeciesGeneMat.Bl function
data(Seq.DF4) ## the first object of the list is the Species-by-gene matrix

# Run the function without a pre-selection of gene regions.
SelGene.MaxSpCov(input = Seq.DF4[[1]])
# Run the function with a pre-selection of gene regions.
SelGene.MaxSpCov(input = Seq.DF4[[1]], NBGene = c("co1", "12srrna"))
SpbyGeneMat=rbind(as.matrix(Seq.DF4[[1]]), c("Titi_titi",0, 2, 1), c("Toto_toto", 0, 0, 4))
row.names(SpbyGeneMat)=SpbyGeneMat[,1]
SpbyGeneMat=as.data.frame(SpbyGeneMat)
SelGene.MaxSpCov(input = SpbyGeneMat, NBGene = c("co1", "12srrna"))
SelGene.MaxSpCov(input = SpbyGeneMat, NBGene = c("co1", "16srrna"))
SelGene.MaxSpCov(input = SpbyGeneMat, NBGene = c("co1"))
SelGene.MaxSpCov(input = SpbyGeneMat, NBGene = c("12srrna"))
SelGene.MaxSpCov(input = SpbyGeneMat, NBGene = c("16srrna"))
SelGene.MaxSpCov(input = SpbyGeneMat, NBGene = 1)
SelGene.MaxSpCov(input = SpbyGeneMat, NBGene = 2)
}
