% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seq.Diastocapen}
\alias{Seq.Diastocapen}
\title{Sequences and metadata extracted from the Barcode of Life Database (BOLD), GenBank (NCBI) and the repository of the New Zealand National 
Fish collection for the species "Diastobranchus capensis".}
\format{a list of 6 tables: 
1) This table contains the raw output (sequences and metadata extracted from the Barcode of Life Database (BOLD) 
for the species "Diastobranchus capensis" the 9/10/2018) of the function GetSeq_BOLD from the regPhylo package.}
\usage{
Seq.Diastocapen
}
\description{
This example provides a list of tables storing sequences and metadata for the species "Diastobranchus capensis". 
Each table correspond to the output table of the following functions: GetSeq_BOLD, GetSeqInfo_NCBI.taxid, 
Congr.NCBI.BOLD.perReposit, GeoCoord.WGS84, GeoCodeName.
}
\keyword{datasets}
