% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{seqle}
\alias{seqle}
\title{Detect suites of consecutive numbers in a vector}
\usage{
seqle(x, incr = 1)
}
\arguments{
\item{x}{the vector of interest with the suite of numbers.}
}
\description{
a function to detect suites of consecutive numbers in a vector from 
https://stackoverflow.com/questions/8400901/group-integer-vector-into-consecutive-runs/8402950#8402950
This function is used internally by the \emph{rm.del.gap} function.
}
