% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Align.Concat.R
\name{Split.Codon.Partition}
\alias{Split.Codon.Partition}
\title{Add the three codons position to the partition file in nexus format}
\usage{
Split.Codon.Partition(
  input.parti = NULL,
  which.codon = NULL,
  output.parti = NULL
)
}
\arguments{
\item{input.parti}{the path including the name of the Partition_file.nex exported 
by the Align.Concat function in nexus format.}

\item{which.codon}{a vector of interger providing the position of the gene in 
the Supermatrix that require to include the three codons positions.
 To know the position of the coding gene you can either look at the print stdout 
 from the Align.Concat function, or upload the "Convtab.txt" table provided by
 the Align.Concat function.}
}
\value{
This function returns a Partition file in enxus format including the 
additional three codon positions for the coding gene
}
\description{
This function use the 'Partition_Concat.nex' file provided by the 
Align.Concat, and provide the additional three codon positions for the coding gene in another
Partition_file in nexus format. This file is ready to be used by IQtree or RAxML
}
