% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxreport2Sp.List.R
\name{Taxreport2Sp.List}
\alias{Taxreport2Sp.List}
\title{Convert a tax_report.txt file from the NCBI taxonomic facility to species lists used by GetSeqInfo_NCBI_taxid and GetSeq_BOLD functions.}
\usage{
Taxreport2Sp.List(input = NULL)
}
\arguments{
\item{input}{path to the "tax_report.txt" file exported by the NCBI taxonomic facility
(\url{https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi})}
}
\value{
The function reports a list of two tables: the first table $SpList.NCBI reports two columns,
the first one reports the NCBI taxid, and the second the binomial species names.
The second table $SpList.BOLD reports two columns, the first one reports all binominal species names
potentially including synonyms used in NCBI taxon database that will also be used as query in BOLD,
and the second column reports the accepted species name that is going to be reported in the output table of \code{\link{GetSeq_BOLD}}.
}
\description{
This function converts the tax_report.txt file exported by the NCBI taxonomic facility
(\url{https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi}) into a list of two tables with
two columns with the species list appropriate to query DNA sequences in GenBank and associated databases
(through NCBI) using the function \code{\link{GetSeqInfo_NCBI_taxid}}, and in BOLD database using the
function \code{\link{GetSeq_BOLD}}.
}
\examples{
\dontrun{
# To run the example copy the input tax_report.txt file provided
# by the regPhylo package to a temporary directory created into the
# current working directory.
src.dir = system.file("extdata/tax_export", package = "regPhylo")
dir.create("TempDir")
# Set up the path to the TempDir folder.
dest.dir = paste(getwd(), "/TempDir", sep="")
file.names <- dir(src.dir)
# Copy all the files stored in regPhylo/extdata/multi.align"
# into a temporary folder.
sapply(file.names, function(x) {
file.copy(from = paste(src.dir, x, sep = "/"),
to = paste(dest.dir, x, sep = "/"),
overwrite = FALSE) })

# Run the function using this example of tax_report.txt file
# for 30 species exported by NCBI taxonomic facility.
taxreport = Taxreport2Sp.List(input = "TempDir/tax_report.txt")
names(taxreport) # the first element of the list is the table
# for NCBI search, the second is for BOLD search.

## Table for NCBI search.
head(taxreport$SpList.NCBI)
dim(taxreport$SpList.NCBI) # one taxa did not have a taxid
# when the request was performed the 7/03/2019.

## Table for BOLD search.
head(taxreport$SpList.BOLD)
dim(taxreport$SpList.BOLD) # two taxa got another preferred NCBI species name.
tail(taxreport$SpList.BOLD)
}
}
