% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect.Outlier.Seq.R
\name{Detect.Outlier.Seq}
\alias{Detect.Outlier.Seq}
\title{Helper to detect outlier sequences in an alignment}
\usage{
Detect.Outlier.Seq(inputal = NULL, Strat.DistMat = NULL,
  Dist.Th = NULL, output = NULL, Second.Outlier = NULL,
  Bitsc.Th = NULL)
}
\arguments{
\item{inputal}{it can be an object of class "alignment" (seqinr R pcakage),
or "DNAbin" (ape R package) or the name (including the path if necessary) of the input fasta file.}

\item{Strat.DistMat}{can be 'MisAli', 'DivSeq' or 'Comb', see details.}

\item{Dist.Th}{distance threshold to detect the primary set of
outlier sequences (between [0-1]).}

\item{output}{name of the output table exported in the working directory.}

\item{Second.Outlier}{if 'TRUE', blasts the primary outlier seqeunces
against a local database based on the sequencs alignment to detect secondary outliers.}

\item{Bitsc.Th}{bitscore similarity threshold to detect 'secondary'
outlier sequences, should be a value between [0-1].}
}
\value{
The function return a table for all the retained sequences with the following headers:
'Query_SeqName', 'SeqLen_Query', 'Hit_SeqName', 'Hit_SeqLen', 'evalue',
'pident', 'length', 'mismatch', 'gapopen', 'qstart', 'qend', 'sstart', 'send',
'bitscore', 'qcovs' Name of the query sequence, length of the query sequence,
name of the hit sequence, length of the hit sequence, E-value, means Percentage
Number of gap openings, means start of alignment in query, means End of
alignment in query, means Start of alignment in subject, means End of alignment
in subject, means Bit score, means Query Coverage Per Subject (for all HSPs)
(See https://www.ncbi.nlm.nih.gov/books/NBK279675/ for more details and
options)
}
\description{
The function helps to relatively quickly detect some outlier sequences (miss
aligned sequences, which can be caused by different problems such as gene or
species annotation problems, presence of paralogs sequences...) that should be
potentially removed from the pool of sequences per species and gene regions
before selecting the best sequence.  The function is far from being accurate
and is only provided as a tool to help detecting potential outlier sequences.
Great care must be taken to the results that may erroneously include outlier
sequences or on the contrary omit outlier sequences. Eyes inspection and
checking must always be carry on to ensure a satisfaying results.
}
\details{
The function follows the following general strategy to detect outlier
sequences.  First a distance matrix is computed among all sequences (See
different strategies available below), second a BIONJs tree (i.e. an improved
Neighbhor joining allowing missing values, Criscuolo & Gascuel (2008, DOI:
10.1186/1471-21.5-9-166) using the disance matrix is build, and the root to tip
distance is computed. All the sequence above a certain distance threshold (we
use 50-60%, see Cheng et al. 2015 for similar thresholds) of the maximun root
to tip distances are extracted as potential 'primary' outlier sequences.  In
option, we can blast the primary outlier sequences on a local blast database to
look for secondary outlier sequences that might be similar to the primary
sequences but just below the distance threshold. All the sequences with a
bitscore above a bitscore threshold (We usually use 80%) obtained when blasting
the query sequence against itself are considered as potential 'secondary'
outliers. We used the bitscore because it is not influenced by the size of the
local database contrary to the E.value
(https://www.ncbi.nlm.nih.gov/BLAST/tutorial/Altschul-1.html).

Three options are available to compute the distance matrix among sequences.
The first otpion 'MisAli' uses the number of indelblocks to compute the
distance matrix in order to preferentially detect misaligned sequences
generating gaps in the alignment.  The second option 'DivSeq' uses the TN93
substitution model to compute the distance matrix in order to detect very
divergent sequences.  The third option 'Comb' combined the two previous
distance matrices to detect both misaligned and very divergent sequence. The
two distance matrix are first re-scaled as percentage of the their highest
distances respectively and then summed.

The function requires BLAST installed and in the PATH, and ape and seqinr R
packages.
}
\examples{
# Load the alignment file (class "alignment"), this is the first
# object of the list called Example_16S_outlier.
data(Example_16S_outlier)
Example_16S_outlier_align = Example_16S_outlier[[1]]
\dontrun{

# Running the function with the 'Comb' option,
# a distance threshold of 0.6, and disabling the search for secondary outlier sequences.
S16_MisAlign0.6_1 = Detect.Outlier.Seq(inputal = Example_16S_outlier_align,
Strat.DistMat = "Comb", Dist.Th = 0.6, output = "Example_16S_outliers_1.txt",
Second.Outlier = "No")
dim(S16_MisAlign0.6_1) ### 36 sequences detected as primary oultier sequences.
head(S16_MisAlign0.6_1)

# The output table is also present as an external data table provided in the regPhylo r package
and can be access by the following code:
# a = system.file("extdata/ExampleOutliers/Example_16S_outliers_1.txt", package = "regPhylo")
# Example_16S_outliers_1 = read.delim(a, sep="\\t", header = TRUE)


# Running the function with the 'Comb' option, a distance threshold of 0.6,
# and allowing a search for secondary outlier sequences using local blast database.
S16_MisAlign0.6_2 = Detect.Outlier.Seq(inputal = Example_16S_outlier_align,
Strat.DistMat = "Comb", Dist.Th = 0.6, output = "Example_16S_outliers_2.txt",
Second.Outlier = "Yes", Bitsc.Th = 0.8)
length(unique(S16_MisAlign0.6_2[,3])) ### 38 primary and secondary outlier seqeunces detected.
head(S16_MisAlign0.6_2)

# The output table is also present as an external data table provided in the regPhylo r package
and can be accessed by the following code:
# a = system.file("extdata/ExampleOutliers/Example_16S_outliers_2.txt", package = "regPhylo")
# Example_16S_outliers_2 = read.delim(a, sep="\\t", header = TRUE)

# To clean the file created while running the example do the following:
file.remove("16S_example_RMoutliers.fas")

}
}
