% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{Detect.stop.codon}
\alias{Detect.stop.codon}
\title{Detect automatically the reading frame and sequence with the presence of stop codon 
in a coding DNA sequence}
\usage{
Detect.stop.codon(input = NULL, genet.code = 5)
}
\arguments{
\item{input}{a "DNAbin" or an "alignment" object containing the multiple DNA sequences already aligned.}

\item{genet.code}{a number precising the genetic code of the DNA sequence. See the help page of the function
\emph{translate} from the \emph{seqinr} R package available at https://cran.r-project.org/web/packages/seqinr/seqinr.pdf}
}
\value{
The functions returns a data.frame with the sequence names including stop codons and the number of stops codon detected.
}
\description{
This function automatically detect the reading frame of a DNA sequences coding for a protein, 
and then detect the sequence with the presence of stop codon.
}
