% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First.Align.All.R
\name{First.Align.All}
\alias{First.Align.All}
\title{Automatically reverse complement the sequence and align the sequences using MAFFT and PASTA}
\usage{
First.Align.All(input = NULL, output = NULL, nthread = NULL,
  methods = NULL, Mafft.path = NULL)
}
\arguments{
\item{input}{path to the folder storing the files of interest, (the files must have
the extension '.fas' and the name of the gene region must be immediately before
the extension separated from the rest of the name by '_', e.g.
'Alignment_co1.fas').}

\item{output}{path to the folder storing the final alignments,
(the output folder is created directly by the function). The name of the
program appears as a prefix in the name of the alignment file using '_' as
separator. The function also exports a table in the input folder with the name
of all the sequences per gene region that have been reverse complemented.}

\item{nthread}{number of threads used to run the alignment software in parallel for
the different gene regions.}

\item{methods}{programs used to aligned the sequences, the function used Mafft-FFTNS1 anyway,
but other programs can be used simultaneously, such as  Mafft-fftnsi and PASTA, c("mafftfftnsi", "pasta").}

\item{Mafft.path}{for Windows plateform, a character string which provides the path
to the mafft executable. For Linux the mafft software must be in the $PATH.}
}
\value{
The function returns the alignment in the output folder using the
name of the program as prefix of the alignment (e.i. "mafftfftnsi_Alig_co1.fas").
The function also returns a table called "ListSeq_RevCompl_FirstAlignAll.txt"
in the output folder listing the seqeunces names that have been reverse
complemented by Mafft.
}
\description{
First, for each distinct gene region, this function aligns all the sequences
using Mafft v7.222 fftns1 (Katoh et al. 2005) with the
option --adjustdirection to automatically detect the direction of the sequences
and reverse complement the sequences if necessary (the function adds a '_R_' as
prefix on the sequence name). Second, the function runs Mafft fftnsi (Katoh et
al. 2005), and PASTA (Mirarab et al. 2015) in parallel and alphabetically orders
the sequences within each alignment (i.e. for each gene region).
}
\details{
PASTA (as Sate I and II) may be better at handling very
divergent sequences and at dealing with saturated phylogenetic signal because
it splits the alignment into small chuncks of similar sequences and performs
profile alignments among those smaller chunk of sequences (Roquet et al. 2013).

The 'output', 'input', and 'nthread' objects need
to be present in the R environment before running the function because the
function runs in parallel using the R package 'parallel', see example below the
function.

The function requires that Mafft and PASTA to be installed and in the PATH.
The function also requires the seqinr and parallel R packages.
}
\examples{
# The input, output and nthread objet have to be present in
# the R global environment before to run the function.
\dontrun{
# We used small alignments examples provided in the
# inst/extdata/FirstToAlign folder available with the package.
# Those alignments are the export of Seq.DF5 dataset, for
# the co1 and 16s regions. Notice that the 16s has only 1
# sequences and explained the warnings messages.

# To run the example it might be better to copy the input alignment files
# provided by the package to a temporary directory created into the
# current working directory.
src.dir = system.file("extdata/FirstToAlign", package = "regPhylo")
dir.create("TempDir.FirstToAlign")
# Set up the path to the TempDir folder.
dest.dir = paste(getwd(), "/TempDir.FirstToAlign", sep="")
file.names <- dir(src.dir)
# Copy all the files stored in regPhylo/extdata/FirstToAlign"
# into a temporary folder.
sapply(file.names, function(x) {
file.copy(from = paste(src.dir, x, sep = "/"),
to = paste(dest.dir, x, sep = "/"),
overwrite = FALSE) })

# Remove the empty file created for the folder name when
# importing the data file
file.remove("TempDir.FirstToAlign/FirstAligned")

input = "TempDir.FirstToAlign"
output = "TempDir.FirstToAlign/FirstAligned"
nthread = 2
methods = c("mafftfftnsi", "pasta")

First.Align.All(input = input, output =
"TempDir.FirstToAlign/FirstAligned", nthread = 2,
methods = c("mafftfftnsi", "pasta"))


# To clean the file created while running the example do the following:
unlink("TempDir.FirstToAlign", recursive = TRUE)

}
}
\references{
Katoh et al. 2005, DOI: 10.1093/nar/gki198

Mirarab et al. 2015 DOI: 10.1089/cmb.1014.0156

Roquet et al. 2013 DOI: 10.1111/j.1600-0587.2012.07773.x
}
