% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoCoord.WGS84.R
\name{GeoCoord.WGS84}
\alias{GeoCoord.WGS84}
\title{Extract longitude and latitude coordinates in separate fields and conversion in WGS84 using +- North, and +- East}
\usage{
GeoCoord.WGS84(input = NULL, output = NULL)
}
\arguments{
\item{input}{table coming from the function Congr.NCBI.BOLD.perReposit()}

\item{output}{name of the output table exported in the working directory.}
}
\value{
Two identical tables are exported, one in the R environment and one in the working directory.
These tables are similar to the input table except that they have one additional column,
the previous "Lat_Lon" corresponds now to two distinct columns "Latitude" and "Longitude".
}
\description{
This function splits the field 'Lat_Lon' in two distinct fields for 'Latitude'
and 'Longitude' and transform them to WGS84 geographic coordinates using +- North, and +- East (i.e. -45.10
-53.23)
}
\details{
This function uses some code from the '05.extract_coords_from_GB.r' available at
'https://github.com/paolo-gratton/Gratton_et_al_JBiogeogr_2016/blob/master/05.extract_coords_from_GB.r'
and used in the paper Gratton et al. 2016 'A world of sequences: can we use
georeferenced nucleotide databases for a robust automated phylogeography?'
Journal of Biogeography. DOI: 10.1111/jbi.12786.
}
\examples{
# Import a tables exported by the function Congr.NCBI.BOLD.perReposit
data(Seq.Diastocapen)
Seq.DF = Seq.Diastocapen$Seq.DF

# Run the function
Seq.DF1=GeoCoord.WGS84(input=Seq.DF, output="Seq.DF.txt")
dim(Seq.DF1)
}
