% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSeqInfo_NCBI_taxid.R
\name{GetSeqInfo_NCBI_taxid}
\alias{GetSeqInfo_NCBI_taxid}
\title{Extract DNA sequences and metadata from GenBank}
\usage{
GetSeqInfo_NCBI_taxid(splist = NULL, gene = NULL, filename = NULL,
  timeout = 10)
}
\arguments{
\item{splist}{a two column table with the unique NCBI taxonomic ID (i.e. taxid) in
the first column and the accepted binomial species name in the second column
which will be retained as 'TaxaName' in the output file.}

\item{gene}{can be a
particular gene region or 'ALL' if all gene regions are
sought.}

\item{filename}{name of the output table (also needs to include the
extension, e.g. ".txt")}

\item{timeout}{the timeout in seconds for socketConnection, for the
choosebank function of the seqinr R package. Default 10 seconds.
It might be necessary to increase the timeout (time to get answer from
the server) if the function cannot retrieve any DNA sequence for
certain species while DNA sequences are available in GenBank for
these species.}
}
\value{
This function returns a table with the species names and the number of
occurrences retrieved from GenBank in the R environment, and exports a table
with the DNA sequences and the metadata into the working directory.

The table exported in the working directory has the following
 headers:
\itemize{
\item 'TaxaName' (Binomial species name included in the second column of the NCBI
species list),
\item 'AccessNb' (NCBI accession number),
\item 'Sequence' (The sequence itself, or 'Too_Long' if more than 5000bp),
\item 'SeqLength' (Sequence length),
\item 'Definition' (NCBI definition field),
\item 'OrganismClassif' (Classification of the
organism),
\item 'Source' (NCBI source),
\item 'Title' (Title provided by NCBI),
\item 'Authors' (Authors provided by NCBI),
\item 'Journal' (Journal provided by NCBI),
\item 'Pubmed' (Pubmed references provided by NCBI),
\item 'Year' (Year provided by NCBI),
\item 'Organism' (Organism provided by NCBI),
\item 'Organelle' (Organelle provided by NCBI),
\item 'Mol_type' (Mol_type provided by NCBI),
\item 'Db_xref' (Db_xref for NCBI),
\item 'Product' (Product provided by NCBI),
\item 'Genes' (Gene provided by NCBI),
\item 'Location' (Field Country in NCBI),
\item 'isolation_source' (isolation_source provided by NCBI),
\item 'Lat_lon' (Lat_lon provided by NCBI),
\item 'Collection_date' (Collection_date provided by NCBI),
\item 'Date_Extract' (Date of the extraction for the data)
}
}
\description{
This function extracts all the sequences, accession numbers, and related
information for all the gene regions or a given gene region, from a list of species using
their NCBI taxid.
}
\details{
This function extracts data from an external server and can be very slow.
}
\examples{
# A table with two species and their unique NCBI taxa ID
Splist=cbind(TaxID=c(443778,189923),
Species.Name=c("Diastobranchus capensis", "Synaphobranchus affinis"))
# Run the function to extract all DNA sequences and associated metadata
\dontrun{
NCBI.output = GetSeqInfo_NCBI_taxid(splist = Splist, gene = "ALL",
filename = "output.NCBI.txt")

# The output can be loaded by doing the following
data(Seq.Diastocapen)
NCBI.output = Seq.Diastocapen$Seq.NCBI

# To remove the file created while running the example do the following:
file.remove("output.NCBI.txt")

}
}
