% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi.Align.R
\name{Multi.Align}
\alias{Multi.Align}
\title{Align sequences using multiple alignment programs}
\usage{
Multi.Align(input = NULL, output = NULL, nthread = NULL,
  methods = NULL, Mafft.path = NULL, Muscle.path = NULL,
  Prank.path = NULL)
}
\arguments{
\item{input}{path to the folder storing the files of interest for further alignment.
The files should have the extension '.fas' and the name of the gene region
should be immediately before the extension separated from the rest of the name
by '_' (e.g.  'Alignment_co1.fas').}

\item{output}{path to the folder storing the
final alignments. The output folder is created directly by the function. The
name of the program appears as a prefix in the name of the alignment file using
'_' as a separator. The function also exports a table into the input folder containing
all the names of the sequences, per gene region, that have been reverse
complemented. A similar table is exported into the R environment.}

\item{nthread}{number of threads used to run the alignment software in
parallel for the different gene regions.}

\item{methods}{programs used to align the sequences. The function uses Mafft-FFTNS1
to reverse complement and quickly align the sequences,
but other programs can be used simultaneously, such as Mafft-fftns2, Mafft-fftnsi,
Muscle, Prank and PASTA, c("mafftfftns2", "mafftfftnsi", "muscle", "prank", "pasta").}

\item{Mafft.path}{for the Windows platform, a character string which provides the path
to the Mafft executable
(e.g. Mafft.path = "C:/Users/deme/Documents/Programs/Mafft/
mafft-7.409-win64-signed/mafft-win/mafft") (see examples below). Even if Mafft is not
the alignment you would like to use, it is necessary to use mafft to reverse complement the sequences
and therefore the Mafft.path must be completed in all cases.
For Linux the mafft software must be in the $PATH.}

\item{Muscle.path}{for the Windows platform, a character string which provides the path
to the Muscle executable (e.g. "C:/Users/deme/Documents/Programs/Muscle/muscle3.8.31_i86win32.exe").
For Linux the Muscle software must be in the $PATH.}

\item{Prank.path}{for the Windows platform, a character string which provides the path
to the Prank executable
(e.g. "C:/Users/deme/Documents/Programs/Prank/prank.windows.140603/prank/bin/prank.exe")
For Linux the Prank software must be in the $PATH.}
}
\value{
This function returns the new alignments into the output folder.
The output folder is created directly by the function. The
name of the program appears as a prefix in the name of the alignment file using
'_' as a separator. The function also exports a table into the input folder containing
all the names of the sequences, per gene region, that have been reverse
complemented, the same table is exported into the R environment as well.
}
\description{
This function enables running several alignment softwares in parallel, including Muscle v3.8.31 (Edgar 2004),
Mafft v7.222 fftns1, fftns2, fftnsi (Katoh et al. 2005), Prank v.140603 (Loytymoja & Goldman 2008)
and PASTA (Mirarab et al. 2015).
\strong{Note:} PASTA is not available for Windows users.
}
\details{
First, for each distinct gene region, the function aligns all the sequences
using Mafft v7.222 fftns1 (Katoh et al. 2005)
with the option \emph{--adjustdirection} to automatically detect the direction of the
sequences and reverse complement the sequences if necessary (the function adds
an '_R_' as a prefix on the sequence name if it has been reverse complemented).
Second, the function runs Mafft fftns2, fftnsi, Prank and PASTA, and
alphabetically orders the sequences within each of the gene region alignments.

The 'output', 'input' and 'nthread' objects need
to be present in the R environment before running the function
because the function runs in parallel using the R package 'parallel' (see
example).

The function requires that Muscle, Mafft, Prank and Pasta are installed and in
the PATH.
\itemize{ \item For Muscle, download for the different OS are available at
\url{https://www.drive5.com/muscle/downloads.htm},
and instructions at \url{https://www.drive5.com/muscle/manual/}.
\item For Mafft, download for the different OS and all documentation are available at
\url{https://mafft.cbrc.jp/alignment/software/} (for Windows OS we have only tested the
"All-in-one package for Windows" version of Maftt).
\item For Prank, download for the different OS and all documentation are available at
\url{http://wasabiapp.org/software/prank/}, specific instructions for installation are
available at \url{http://wasabiapp.org/software/prank/prank_installation/}.
\item For PASTA, download and instructions for installation are provided on the Github page
available at \url{https://github.com/smirarab/pasta}.
}
}
\examples{
# Load into the R environment the object used by the function.
# Here the example consists of four alignment (co1, 12srrna, cytb, rag1)
# for 16 species.
\dontrun{

# To run the example copy the input alignment files
# provided by the package to a temporary directory created in the
# current working directory.
src.dir = system.file("extdata/multi.align", package = "regPhylo")
dir.create("TempDir.Multi.aligned")
# Set up the path to the TempDir folder.
dest.dir = paste(getwd(), "/TempDir.Multi.aligned", sep="")
file.names <- dir(src.dir)
# Copy all the files stored in regPhylo/extdata/multi.align"
# into a temporary folder.
sapply(file.names, function(x) {
file.copy(from = paste(src.dir, x, sep = "/"),
to = paste(dest.dir, x, sep = "/"),
overwrite = FALSE) })

output = "multi.alignedTest"
nthread = 3
methods = c("mafftfftnsi", "pasta")
input = "TempDir.Multi.aligned"
# run the function using two additional alignment program "mafftfftnsi" and "pasta"
# (not availbale on Windows OS, see the example for Windows below).
Multi.Align(input = input, output = "multi.alignedTest", nthread = 3,
methods = c("mafftfftnsi", "pasta"))
list.files("multi.alignedTest")

# To clean the file created while running the example do the following:
# Remove the temporary folder
unlink("TempDir.Multi.aligned", recursive = TRUE)
# Remove the folder with Gblocks outputs
unlink("multi.alignedTest", recursive = TRUE)


#### To run Multi.Align on Windows OS. ###

output = "multi.alignedTest"
nthread = 3
methods = c("mafftfftnsi", "muscle", "prank")
input = "TempDir.Multi.aligned"

# Example to run "mafftfftnsi", "muscle", and "prank" on Windows.
# (Mafft.path, Muscle.path, Prank.path need to be adapted to the user's
# computer configuration, and are provided hereas an example).

Multi.Align(input = input, output = "multi.alignedTest", nthread = 3,
methods = c("mafftfftnsi", "muscle", "prank"),
Mafft.path = "C:/Users/deme/Documents/Programs/Mafft/mafft-7.409-win64-signed/mafft-win/mafft",
Muscle.path = "C:/Users/deme/Documents/Programs/Muscle/muscle3.8.31_i86win32.exe",
Prank.path = "C:/Users/deme/Documents/Programs/Prank/prank.windows.140603/prank/bin/prank.exe")

#### end specifications to run Multi.Align on Windows OS. ####


# The output files are also present as external data in the regPhylo package
# and can be accessed by the following code:
# a = system.file("extdata/multi.align/multi.aligned", package = "regPhylo")
# list.files(a)
}

}
\references{
Edgar 2004, DOI: 10.1093/nar/gkh340

Katoh et al. 2005, DOI: 10.1093/nar/gki198

Loytymoja & Goldman 2008, DOI: 10.1126/science.1158395

Mirarab et al. 2015, DOI: 10.1089/cmb.1014.0156
}
