% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.GeOMe.XY.R
\name{Query.GeOMe.XY.R}
\alias{Query.GeOMe.XY.R}
\title{Retrieve accurrate geographic coordinates from GeOMe database.}
\usage{
Query.GeOMe.XY.R(input = NULL, Phylum = NULL, output = NULL)
}
\arguments{
\item{input}{is a table coming from the function 'GeoCoord.WGS84.R'.}

\item{Phylum}{is the name of the phylum}

\item{output}{name of the output table exported in the working directory.}
}
\value{
Two identical tables are exported, one in the R environment
and one in the working directory with the output name.
The output is similar to the input table except that the fields
Lat Long have been updated with proper geographic coordinates if sequences
have been found in GeOMe, finally the field 'OriginDatabase' is amended by
'_XY-GeOMe' for the corresponding sequences.
}
\description{
This function queries the GeOMe database https://www.geome-db.org/ (Deck et al.
}
\details{
The function checks for species name synonymies using Worms
(http://www.marinespecies.org/plateform) through the worrms R package
(Chamberlain 2017).

The names of the accepted phylums in GeoMe are available when
downloading a GeOMe database template at the following web page
(https://www.geome-db.org/template, and the list of all the phyla
 accessible in the tab 'List').
}
\examples{
# Import Seq.DF1
#' data(Seq.Diastocapen)
Seq.DF1 = Seq.Diastocapen$Seq.DF1

# Run the function
# Seq.DF2 = Query.GeOMe.XY.R(input = Seq.DF1, Phylum = "Chordata", output = "Seq.DF2.txt")
}
