% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelBestSeq.R
\name{SelBestSeq}
\alias{SelBestSeq}
\title{Select the sequences based on geographic proximity
and/or sequence length, and export sequences (fasta format) and metadata for the selected gene regions}
\usage{
SelBestSeq(
  input = NULL,
  output = NULL,
  RefPoint = NULL,
  perReposit = NULL,
  Alignment = NULL,
  MaxSeq = NULL,
  gene.list = NULL,
  SeqChoice = NULL
)
}
\arguments{
\item{input}{a data table coming from the function \code{\link{Select.DNA}} or
\code{\link{SpeciesGeneMat.Bl}} (the data table should have 29 columns).}

\item{output}{the name of the output table (including the path if necessary,
but the folder must be created first).}

\item{RefPoint}{a vector with geographic coordinates with
longitude and latitude in decimal degrees WGS84 of the geographic reference
point of the focal area. (e.g. the reference point for New Zealand, is 174.7976
degrees longitude, and -41.3355 degrees latitude). If RefPoint is NULL then the selection
based on geographic proximity is disabled, and only the selection based on the length
of the sequence is performed.}

\item{perReposit}{if the name of the repository is provided (i.e. the same name as used in the function
\code{\link{Congr.NCBI.BOLD.perReposit}} option 'perReposit') it implies that several
sequences are coming from a personal repository and the function will use the
ID of the sequence provided by the field 'Db_xref' as a unique ID for the
sequence. The function uses the first element (using ';' as separator) present
in the column 'Db_xref'.}

\item{Alignment}{if Alignment = 'T' then fasta files of the sequences for the selected gene regions are
exported.}

\item{MaxSeq}{maximum number of sequences to be exported for a species
when multiple sequences are available for a gene region
(sequences maximising the spatial optimisation and/or length criteria are exported first).
If MaxSeq = 'ALL', then all the sequences available for that gene region are exported.}

\item{gene.list}{a vector of selected gene regions (gene names have to
be consistent with the header of the table with the suffix '_SpDNA_Mat.txt'
exported by the function \code{\link{SpeciesGeneMat.Bl}}).  For example Cytochrome c
oxydase subunit 1, should have to be written 'co1' and not 'COI' or 'COX1'.}

\item{SeqChoice}{select the longest sequence after removing ambiguous nucleotides
(option SeqChoice = 'Longest') or the sequence with the length closest to the
median sequence length (option SeqChoice = 'Median').}
}
\value{
The function returns a table with all the information about the selected sequences
including 3 additional columns: 'SeqLengthNoIndelsNoAmbig' = Sequence length
after removing the indels and ambiguous nucleotides; 'Dist2RefPoint' = Great circle
distance in kilometers from the RefPoint; 'dist2med' = distance to the median
of the sequence length distribution. The alignment of the gene region are exported in
fasta format.
}
\description{
This function selects and exports the sequences for a list of gene
regions for all the species present in the input data table. The selection of
the sequence for a given species and gene region can be optimised considering the most geographically proximate
sequence to the RefPoint. It can additionally, or alternatively, select a
sequence according to the length of the sequence.
}
\details{
The choice based on the sequence length can be made using the longest sequence (option
SeqChoice = 'Longest') or the sequence with a length closest to the median
of the sequence length distribution for a given species and a given gene region
(option SeqChoice = 'Median') (See Antonelli et al. 2017, Systematic Biology DOI:
10.1093/sysbio/syw066 for a similar approach).  The sequence length is computed
after discarding the potential indels ('-') and the ambiguous nucleotides.
}
\examples{
# Load the data table exported by the function Select.DNA
data(Seq.DF4) # The table is called "Seq.DF5" and constitutes the
# sixth object of the list "Seq.DF4".
Seq.DF5 = Seq.DF4$Seq.DF5

# Run the function and export all the sequences in the alignment for
# each species and gene region.
\dontrun{
Seq.DF6=SelBestSeq(input = Seq.DF5, output = "Alig_Seq.DF5.All",
RefPoint = cbind(174.7976, -41.3355), perReposit = "My.Rep",
Alignment = T, MaxSeq = "ALL", gene.list = c("co1", "16srrna"),
SeqChoice = "Median")

dim(Seq.DF6) # 10 sequences are reported.


# Run the function and export the best (the most proximal to the focal
# area i.e. NZ) sequences in the alignment for each species and gene region,
# selecting the sequence with a median sequence length.
Seq.DF7=SelBestSeq(input = Seq.DF5, output = "Alig_Seq.DF5.Best",
RefPoint = cbind(174.7976, -41.3355), perReposit = "My.Rep",
Alignment = T, MaxSeq =1, gene.list = c("co1", "16srrna"),
SeqChoice = "Median")

dim(Seq.DF7) # 2 sequences are reported.


# Run the function and export the two most proximal sequences of the focal area (i.e. NZ)
# in the alignment for each species and gene region, selecting the longest sequence.
Seq.DF8=SelBestSeq(input = Seq.DF5, output = "Alig_Seq.DF5.Best",
RefPoint = cbind(174.7976, -41.3355), perReposit = "My.Rep",
Alignment = T, MaxSeq =2, gene.list = c("co1", "16srrna"),
SeqChoice = "Longest")

dim(Seq.DF8) # 3 sequences are reported: 2 CO1 sequences, and the only 16srrna
# sequence available.


# Run the function to select the two longest sequences per gene region and species,
# regardless of geographic proximity.
Seq.DF9=SelBestSeq(input = Seq.DF5, output = "Alig_Seq.DF5.Best",
perReposit = "My.Rep", Alignment = T, MaxSeq = 2,
gene.list = c("co1", "16srrna"), SeqChoice = "Longest")

dim(Seq.DF9) # 3 sequences are reported: 2 CO1 sequences and the only 16srrna
# sequence available.
}

}
\references{
Antonelli et al. 2017, DOI: 10.1093/sysbio/syw066
}
