% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.DNA.R
\name{Select.DNA}
\alias{Select.DNA}
\title{Extract all the seqeunces and metadata from the data table for a list of selected gene regions}
\usage{
Select.DNA(input = NULL, gene.list = NULL, output = NULL)
}
\arguments{
\item{input}{The table created using the function \code{\link{SpeciesGeneMat.Bl}}
with the suffix '_CleanDataset.txt'.}

\item{gene.list}{a vector of the selected gene regions (gene names have
to be consistent with the header of the table with the suffix '_SpDNA_Mat.txt'
exported by the function SpeciesGeneMat_Bl.R). For example (Cytochrome c oxidase subunit 1, should be
written 'co1' and not 'COI' or 'COX1').}

\item{output}{the name of the output table exported into the working directory.}
}
\description{
This function selects the sequences and associated information for a list
of selected gene regions for all the species present in the data table. For
full mitochondrial genomes, or all sequences with a sequence length > 5000 bp
listed as 'TooLong', the function extracts the sequences regions of interest (from GenBank)
 using the query function of the seqinr R package.
}
\examples{
# Load the data table exported by the SpeciesGeneMat.Bl function
data(Seq.DF4) # the table of sequences and metadata is called "CleanDataTable"
# and is the fifth object of the list "Seq.DF4".

# Run the function
Seq.DF5=Select.DNA(input = Seq.DF4$CleanDataTable, gene.list = c("co1", "16srrna"),
output = "Seq.DF4.dataTable")

}
