% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seq.DF4}
\alias{Seq.DF4}
\title{Output of the function SpeciesGeneMat.Bl and Select.DNA for the species "Diastobranchus capensis"}
\format{A list of 5 data frames and one string, 
1) The first table provides the species-by-gene matrix}
\usage{
Seq.DF4
}
\description{
This output is a list of 6 objects: the first five element are the output of the SpeciesGeneMat.Bl function, the sixth object is the table output of the Select.DNA function.
The first table provides the species-by-gene matrix, the second table provides a summary table from a gene perspective (number of species and sequences for each gene region), and the third table provides a summary tables from the species perspective (number of gene regions and sequences for each species), the fourth object provides an empty string listing the species lost when removing microsatellites, unassigned DNA, or blacklisted sequences, the fifth element is a table including all the sequences and the metadata exported by the function SpeciesGeneMat.Bl after removing microsatellites, unassigned DNA and potentially blacklisted sequences. This table also includes two additional columns, one reporting the entry order of the sequences, and the other the name of the gene region used for the species-by-gene matrix.
Finally the last table (i.e. "Seq.DF5") ouptut of the
}
\keyword{datasets}
