% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.Seq.R
\name{seqle}
\alias{seqle}
\title{for Debug
rm(Remove.Seq.TooManyAmbig, Remove.Short.Seq, Remove.Seq.Indels, Seq.removed, Seq.removed.DF, output, NewAlign.Noindels2, Seq.names.final, Seq.removed, Seq.To.remove.short, Seq.to.remove.indels, Seq.To.remove.amb, Seq.To.remove, nbAmbig,nb.loc, Seq.To.remove.Too.Ambig, nbmissing, cytb2Smat, bb, Pos.ToRemove, resa, Ending, Starting, seq.name.indel, Test.Ending, Test.preceding, Seq.to.remove.Indel, aa, nbindels)}
\usage{
seqle(x, incr = 1)
}
\description{
a function to detect sequences of consecutive numbers from https://stackoverflow.com/questions/8400901/group-integer-vector-into-consecutive-runs/8402950#8402950
}
