% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxreport2Sp.List.R
\name{Taxreport2Sp.List}
\alias{Taxreport2Sp.List}
\title{Convert a tax_report.txt file from the NCBI taxonomic facility to species lists used for GetSeqInfo_NCBI_taxid and GetSeq_BOLD functions.}
\usage{
Taxreport2Sp.List(input = NULL)
}
\arguments{
\item{input}{path to the tax_report.txt file exported by the NCBI taxonomic facility (https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi)}
}
\description{
This function converts the tax_report.txt file exported by the NCBI taxonomic facility 
(https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi) into a list of two tables with 
two columns with the species list appropriate to query DNA sequences in GenBank and associated databases 
(through NCBI) using the function GetSeqInfo_NCBI_taxid, and in BOLD database using the function GetSeq_BOLD.
}
