% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxreport2Sp.List.R
\name{Taxreport2Sp.List}
\alias{Taxreport2Sp.List}
\title{Convert a tax_report.txt file from the NCBI taxonomic facility to species lists used by GetSeqInfo_NCBI_taxid and GetSeq_BOLD functions.}
\usage{
Taxreport2Sp.List(input = NULL)
}
\arguments{
\item{input}{path to the "tax_report.txt" file exported by the NCBI taxonomic facility
(\url{https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi})}
}
\value{
This function creates a list of two tables: the first table $SpList.NCBI has two columns,
the first one reports the NCBI taxid, and the second column contains the binomial species names.
The second table $SpList.BOLD also has two columns, the first column reports all binominal species names,
including synonyms used in NCBI taxon database that will be used to query BOLD,
and the second column reports the accepted species name that is going to be reported in the output table of \code{\link{GetSeq_BOLD}}.
Note: notice that in table $SpList.BOLD used to query BOLD, additional synonyms can be further added manually to those identified
in the NCBI taxon database.
}
\description{
This function converts the tax_report.txt file exported by the NCBI taxonomic facility
(\url{https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi}) into a list of two tables containing
the species lists to query DNA sequences in GenBank and associated databases
(through NCBI using the function \code{\link{GetSeqInfo_NCBI_taxid}}), and the BOLD database (using the
function \code{\link{GetSeq_BOLD}}).
}
\examples{
\dontrun{
# To run the example, copy the input tax_report.txt file provided
# by the regPhylo package to a temporary directory created in the
# working directory.
src.dir = system.file("extdata/tax_export", package = "regPhylo")
dir.create("TempDir")
# Set up the path to the TempDir folder.
dest.dir = paste(getwd(), "/TempDir", sep="")
file.names <- dir(src.dir)
# Copy all the files stored in regPhylo/extdata/multi.align"
# into a temporary folder.
sapply(file.names, function(x) {
file.copy(from = paste(src.dir, x, sep = "/"),
to = paste(dest.dir, x, sep = "/"),
overwrite = FALSE) })

# Run the function using this example of tax_report.txt file
# for 30 species exported by the NCBI taxonomic facility.
taxreport = Taxreport2Sp.List(input = "TempDir/tax_report.txt")
names(taxreport) # the first element of the list is the table
# for the NCBI search, the second is for the BOLD search.

## Table for NCBI search.
head(taxreport$SpList.NCBI)
dim(taxreport$SpList.NCBI) # one taxa did not have a taxid
# when the request was performed on the 7/03/2019.

## Table for BOLD search.
head(taxreport$SpList.BOLD)
dim(taxreport$SpList.BOLD) # two taxa have another preferred NCBI species name.
tail(taxreport$SpList.BOLD)
}
}
